added tests for unix sockets and fixed the issue with start_unix_server

+++ src/server/mod.rs
@@ -21,7 +21,10 @@ pub async fn start_server(addr: String) -> Result<(), tokio::io::Error> {
 }
 
 pub async fn start_unix_server(path: String) -> Result<(), tokio::io::Error> {
-    std::fs::remove_file(path.clone())?;
+    if std::path::Path::new(&path).exists() {
+        std::fs::remove_file(path.clone())?;
+    }
+
     let listener = UnixListener::bind(&path)?;
     info!("Listening on: {}", path);
     info!("getting global broadcaster");
